requirements.txt
google-generativeai
playwright
psycopg2-binary
sqlalchemy
python-dotenv
fastapi
uvicorn
pydantic
openai
